% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CareerStats.R
\name{CareerStats}
\alias{CareerStats}
\title{Gather stats over the entire year range of a table}
\usage{
CareerStats(data, agg = sum)
}
\arguments{
\item{data}{The table to aggregate stats across years for}

\item{agg}{The aggregate function to apply to numeric columns, Default: sum()}
}
\value{
A dataframe, the original table grouped by player with the aggregate function applied to numeric columns
}
\description{
Gather stats over the entire year range of a table
}
\details{
Also adds to the return table a column for debut year, retirement year, and years active (all within the limits of data).
data must include playerID and yearID to function.
}
\examples{
# The Yankees players with the best career OPS since integration
CareerStats(
  battingStats(
    QuickStats(Batting, teams = "NYA", years = eraToYears("integration", "modern")) \%>\%
    filter(G>100)
  ),
  agg = mean
) \%>\%
select(playerID, G, BA:yearsActive) \%>\%
arrange(-OPS)
}
