% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eraToYears.R
\name{eraToYears}
\alias{eraToYears}
\title{Convert historic era terms to a range of years}
\usage{
eraToYears(startEra, endEra = startEra)
}
\arguments{
\item{startEra}{The era to begin the year range on}

\item{endEra}{The era to go through and end the year range on, Default: startEra}
}
\value{
A vector of two years, the start and end of a range within 1871-2023
}
\description{
Accepts a named era in baseball history and returns the
range of years covering that era (as used on baseball-reference.com).
The acceptable eras and the years they cover are:\enumerate{
\item "preMLB": 1871-1903
\item "deadBall": 1901-1919
\item "liveBall": 1920-1941
\item "integration": 1942-1960
\item "expansion": 1961-1976
\item "freeAgency": 1977-1993
\item "steroid": 1994-2005
\item "modern": 2006-2023}
If an argument is provided for endEra, this will return the first year in startEra
and the last year of endEra. Return the year range of only the given startEra by default.
}
\examples{
eraToYears("liveBall", endEra = "freeAgency")
}
