% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerID.R
\name{getPlayerID}
\alias{getPlayerID}
\title{Quickly convert between a player's name and his code in the Lahman database}
\usage{
getPlayerID(name)
}
\arguments{
\item{name}{The player's full name, "First Last"}
}
\value{
A string, the playerID for the player with the given name. If there are multiple matching players, this will be a vector of strings
}
\description{
If there are more than one matching players, this will return all of them.
Typically, the numeric suffix for these players is in ascending order by debut date.
}
\examples{
getPlayerID("Dansby Swanson")
getPlayerID("Ken Griffey")
}
