% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerName.R
\name{getPlayerName}
\alias{getPlayerName}
\title{Get a player's name from his playerID}
\usage{
getPlayerName(playerID)
}
\arguments{
\item{playerID}{The player's ID in the Lahman::People table}
}
\value{
A string, the player's name ("First Last")
}
\description{
Get a player's name from his playerID
}
\examples{
getPlayerName("johnsra05")
}
