% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTeamID.R
\name{getTeamID}
\alias{getTeamID}
\title{Get the database ID for a team}
\usage{
getTeamID(teamName)
}
\arguments{
\item{teamName}{The team to get the ID for, "City TeamName"}
}
\value{
A string, that team's three letter abbreviation
}
\description{
Get the database ID for a team
}
\examples{
getTeamID("chicago cubs")
}
