% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTeamName.R
\name{getTeamName}
\alias{getTeamName}
\title{Get a team's name from its three letter abbreviation}
\usage{
getTeamName(teamID, most_recent = TRUE)
}
\arguments{
\item{teamID}{The team's ID as used in Lahman::Teams}

\item{most_recent}{A boolean indicating whether to only give the most recent name used by that team code, Default: TRUE}
}
\value{
A string (the team's name) or a vector of strings (all unique team names corresponding to that code over the years)
}
\description{
Get a team's name from its three letter abbreviation
}
\examples{
getTeamName("alt")
getTeamName("hou", FALSE)
}
