% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerPos.R
\name{playerPos}
\alias{playerPos}
\title{Get the fielding position(s) for a given player}
\usage{
playerPos(pID, list = FALSE, years = c(1871, 2023))
}
\arguments{
\item{pID}{The playerID of the player to search}

\item{list}{A boolean; if TRUE, function will return a tibble for the player indicating all positions played and the number of games at each position. If FALSE, just returns the most frequently played position, Default: FALSE}

\item{years}{A single year or a vector of (start year, end year) indicating what years to include for the given player}
}
\value{
A string (the player's primary position) or a tibble with POS and totalG columns
}
\description{
Get the fielding position(s) for a given player
}
\examples{
playerPos(getPlayerID("sammy sosa"))
playerPos(getPlayerID("dylan moore"), list=T, years=2019)
}
