% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerTeam.R
\name{playerTeam}
\alias{playerTeam}
\title{Get the team(s) a player played for}
\usage{
playerTeam(pID, list = FALSE, years = c(1871, 2023))
}
\arguments{
\item{pID}{The player's playerID}

\item{list}{A boolean; if TRUE, function will return a tibble for the player indicating all teams played on and the number of seasons with each. If FALSE, just returns the team played at longest, Default: FALSE}

\item{years}{A single year or a vector of (start year, end year) indicating what years to include for the given player}
}
\value{
A string (the player's primary position) or a tibble with team, seasons, first year, and last year columns
}
\description{
Get the team(s) a player played for
}
\examples{
playerTeam("ruthba01", list = T)

playerTeam(getPlayerID("hank aaron"), years = c(1965, 1970))
}
