% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickStats.R
\name{QuickStats}
\alias{QuickStats}
\title{Quickly filter a table from the Lahman database}
\usage{
QuickStats(
  data,
  players = unique(data$playerID),
  years = c(1871, 2023),
  teams = unique(data$teamID),
  lg = unique(data$lgID),
  pos = c("P", "LF", "CF", "RF", "2B", "3B", "1B", "SS", "C"),
  idvars = colnames(data)
)
}
\arguments{
\item{data}{One of the tables from Lahman to filter for analysis}

\item{players}{A vector of or a single playerID, the players to include. Default: All in data}

\item{years}{A single year or a vector of (start year, end year) indicating what years to include, Default: c(1871, 2023)}

\item{teams}{A vector of or a single teamID, the teams to include. Default: All in data}

\item{lg}{A vector of or a single lgID, the leagues to include. Default: All in data}

\item{pos}{A vector of or a single position abbreviation, the positions to include. Default: All positions}

\item{idvars}{The columns to return, Default: all columns in data + POS}
}
\value{
A dataframe, the filtered table
}
\description{
Quickly filter a table from the Lahman database
}
\examples{
# postseason batting stats for National League centerfielders during the steroid era
QuickStats(BattingPost, years=eraToYears("steroid"), lg="NL", pos = "CF")

# Cubs pitching stats in 2016
QuickStats(Pitching, teams=getTeamID("chicago cubs"), years = 2016)
}
